% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drone_createLitchiMissionGrid.R
\name{computeRotationMatrix}
\alias{computeRotationMatrix}
\title{Compute a rotation matrix for a given angle}
\usage{
computeRotationMatrix(angle_rad)
}
\arguments{
\item{angle_rad}{The angle in radians for which to compute the rotation matrix.}
}
\value{
A rotation matrix.
}
\description{
Compute a rotation matrix for a given angle
}
\keyword{internal}
