% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drone_createLitchiMissionGrid.R
\name{fp_litchi_mission}
\alias{fp_litchi_mission}
\title{Generate a mission plan for a drone in Litchi app}
\usage{
fp_litchi_mission(photo_grid, origin_lat, origin_long, angleDeg = 0, ...)
}
\arguments{
\item{photo_grid}{A photo grid data frame created using the \code{fp_photo_grid} function.}

\item{origin_lat}{The latitude of the starting point for the mission (in degrees).}

\item{origin_long}{The longitude of the starting point for the mission (in degrees).}

\item{angleDeg}{The angle at which the survey grid is rotated from North (in degrees). e.g. a value of 45 rotates the photo grid around the origin point (bottom-left) by 45 degrees.}

\item{...}{additional arguments to be passed to the function.}
}
\value{
A data frame containing the mission plan for the drone, in a format suitable for use with the Litchi app.
}
\description{
Generate a mission plan for a drone in Litchi app
}
\examples{
photo_grid <- fp_photo_grid(altitude = 60, overlap_width = 0.6,
overlap_height = 0.2, survey_xaxis = 100, survey_yaxis = 200, plot = FALSE)

litchiMission <- fp_litchi_mission(photo_grid, pg, origin_lat = 55.125505,
origin_long = 10.268467, angle = 38)
\dontrun{
write.csv(x = litchiMission, file = "missionFiles/testGrid10.csv",
row.names = FALSE)
}

}
