% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drone_photoGrid.R
\name{fp_photo_grid}
\alias{fp_photo_grid}
\title{Create a grid of points representing the centers of photos}
\usage{
fp_photo_grid(
  altitude = 40,
  overlap_width = 0.4,
  overlap_height = 0.1,
  survey_xaxis = 200,
  survey_yaxis = 300,
  plot = TRUE,
  fov = 94,
  aspectRatio = c(4, 3),
  ...
)
}
\arguments{
\item{altitude}{The altitude at which the photos will be taken (in meters).}

\item{overlap_width}{The width-wise overlap between photos (as a fraction between 0 and 1).}

\item{overlap_height}{The height-wise overlap between photos (as a fraction between 0 and 1).}

\item{survey_xaxis}{The width of the survey area (in meters).}

\item{survey_yaxis}{The height of the survey area (in meters).}

\item{plot}{A boolean value indicating whether to plot the photo grid (TRUE) or return the data frame of points (FALSE).}

\item{fov}{numeric, optional. Field of view of the camera in degrees. Default is 94.}

\item{aspectRatio}{numeric vector, optional. Aspect ratio of the image in the form of a 2-element vector, with the width as the first element and the height as the second element. Default is c(4,3).}

\item{...}{additional arguments to be passed to the function.}
}
\value{
If \code{plot} is TRUE, returns a \code{ggplot} object visualizing the photo grid. If \code{plot} is FALSE, returns a data frame of consisting of photoID and points representing the centers of each photo, and altitude.
}
\description{
Create a grid of points representing the centers of photos
}
\examples{
fp_photo_grid()
fp_photo_grid(
  altitude = 60, overlap_width = 0.6,
  overlap_height = 0.2, survey_xaxis = 400, survey_yaxis = 500, plot = TRUE
)

}
