% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drone_createLitchiMissionGrid.R
\name{transformSf}
\alias{transformSf}
\title{Transform an sf object to a meter-based CRS}
\usage{
transformSf(data)
}
\arguments{
\item{data}{An sf object.}
}
\value{
The input sf object transformed to a meter-based coordinate reference system (CRS).
}
\description{
Transform an sf object to a meter-based CRS
}
\examples{
data <- data.frame(latitude = 45.5236, longitude = -122.675)
data <- convertToSf(data)
transformSf(data)
}
\keyword{internal}
