% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drone_footprint_calculations.R
\name{footprint_}
\alias{footprint_}
\alias{footprint_Diagonal}
\alias{footprint_Height}
\alias{footprint_Width}
\title{Drone Footprint Calculations}
\usage{
footprint_Diagonal(altitude, fov = 94, aspectRatio = c(4, 3), ...)

footprint_Height(altitude, fov = 94, aspectRatio = c(4, 3), ...)

footprint_Width(altitude, fov = 94, aspectRatio = c(4, 3), ...)
}
\arguments{
\item{altitude}{numeric, required. Altitude of the drone in meters.}

\item{fov}{numeric, optional. Field of view of the camera in degrees. Default is 94.}

\item{aspectRatio}{numeric vector, optional. Aspect ratio of the image in the form of a 2-element vector, with the width as the first element and the height as the second element. Default is c(4,3).}

\item{...}{additional arguments to be passed to the function.}
}
\value{
A single numeric value representing the calculated aspect of the camera footprint.
}
\description{
Calculates various aspects of a drone's camera footprint at a given altitude, field of view, and image aspect ratio.
}
\examples{
\dontrun{
# Calculate the width of the camera footprint at an altitude of 100 meters
footprint_Width(altitude = 100)

# Calculate the height of the camera footprint at an altitude of 50 meters
# and a field of view of 120 degrees
footprint_Height(altitude = 50, fov = 120)

# Calculate the diagonal length of the camera footprint at an altitude of
# 200 meters, a field of view of 90 degrees, and an image aspect ratio of
# 4:3
footprint_Diagonal(altitude = 200, fov = 90, aspectRatio = c(4, 3))
}
}
